pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
Flask = "^2.2.0"
urllib3 = "^1.26.12"
bs4 = "^0.0.1"
openai = "^0.26.1"
ratelimiter = "^1.2.0"
sympy = "^1.11.1"
python-status = "^1.0.1"
schedule = "^1.1.0"
qrcode = "^7.3.1"
Pillow = "^9.4.0"
google-api-python-client = "^2.74.0"
pytube = "^12.1.2"
pandas = "^1.5.3"
openpyxl = "^3.1.0"
pdf2image = "^1.16.2"
Wikipedia-API = "^0.5.8"
website = "^1.1"
matplotlib = "^3.6.3"
anime-api = "^0.15.4"
"waifuim.py" = "^3.3.0"
replit = "^3.2.5"
discord = "^2.1.0"
"discord.py" = "^2.1.1"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"